moba_analysis_app/backend/requirements.txt
yt_dlp
fastapi
pydantic
opencv-python
ffmpeg
uvicorn
requests
pandas
python-multipart
mwrogue
mwclient
plotly
Pillow
bs4
kaleido
openpyxl
pytesseract